+++ PxWeb.UnitTests/SavedQuery/SavedQueryDatabaseStorageOptionsTests.cs
@@ -1,10 +1,4 @@
-﻿using System;
-
-using Microsoft.Extensions.Options;
-
-using PxWeb.Code.Api2.SavedQueryBackend.DatabaseBackend;
-
-namespace PxWeb.UnitTests.SavedQuery
+﻿namespace PxWeb.UnitTests.SavedQuery
 {
     [TestClass]
     public class SavedQueryDatabaseStorageOptionsTests
@@ -22,62 +16,7 @@ namespace PxWeb.UnitTests.SavedQuery
             Assert.AreEqual("Microsoft", options.DatabaseVendor);
         }
 
-        [TestMethod]
-        public void Constructor_ShouldThrowArgumentException_WhenDatabaseVendorIsNotSupported()
-        {
-            // Arrange
-            var options = new SavedQueryDatabaseStorageOptions
-            {
-                DatabaseVendor = "UnsupportedVendor"
-            };
-
-            var mockStorageOptions = new Mock<IOptions<SavedQueryDatabaseStorageOptions>>();
-            mockStorageOptions.Setup(o => o.Value).Returns(options);
 
 
-            DataSourceOptions dataSourceOptions = new DataSourceOptions
-            {
-                DataSourceType = "Database"
-            };
-
-            var mockDataSourceOptions = new Mock<IOptions<DataSourceOptions>>();
-            mockDataSourceOptions.Setup(o => o.Value).Returns(dataSourceOptions);
-
-            var resolver = new Mock<ITablePathResolver>();
-
-
-            // Act & Assert
-            Assert.ThrowsException<ArgumentException>(() => new SavedQueryDatabaseStorageBackend(mockDataSourceOptions.Object, mockStorageOptions.Object, resolver.Object));
-        }
-
-
-
-        [TestMethod]
-        public void Constructor_ShouldNotThrow_WhenDatabaseVendorIsSupported()
-        {
-            // Arrange
-            var options = new SavedQueryDatabaseStorageOptions
-            {
-                DatabaseVendor = "Oracle"
-            };
-
-            var mockStorageOptions = new Mock<IOptions<SavedQueryDatabaseStorageOptions>>();
-            mockStorageOptions.Setup(o => o.Value).Returns(options);
-
-
-            DataSourceOptions dataSourceOptions = new DataSourceOptions
-            {
-                DataSourceType = "Database"
-            };
-
-            var mockDataSourceOptions = new Mock<IOptions<DataSourceOptions>>();
-            mockDataSourceOptions.Setup(o => o.Value).Returns(dataSourceOptions);
-
-            var resolver = new Mock<ITablePathResolver>();
-
-            // Act & Assert
-            var backend = new SavedQueryDatabaseStorageBackend(mockDataSourceOptions.Object, mockStorageOptions.Object, resolver.Object);
-            Assert.IsNotNull(backend);
-        }
     }
 }
